add guest resume, ELF64 loader, and fix LLP64 truncations
Implements M3.5 guest resume and R1-R2 rungs of M4. Guest now resumes
correctly across 64 entries, verified on M92p with dual counters. ELF
loader is now arch-conditional with Elf32/Elf64 support and
class/machine validation to prevent silent misparsing. Fixed three LLP64
address truncations: vm_ptr_t width (new patch), interrupt.c ISR
dispatch, passage.c debug print. Added --probe-sweep to map host kernel
PML4 occupancy (Windows XP x64 uses 9 of 256 slots at 493+). Documented
kernel config requirements, measured address-space layout showing no
collisions in the free 118 TB block, and settled M4 to use modern 7.1.5
kernel rather than backporting 2.6.33.

+++ src/colinux/arch/x86_64/interrupt.c
@@ -5,6 +5,9 @@
  * the root directory.
  */
 
+/* GCC's own freestanding stddef.h, for size_t as a pointer-sized integer. */
+#include <stddef.h>
+
 #include <colinux/arch/interrupt.h>
 
 /*
@@ -83,7 +86,12 @@ void co_monitor_arch_real_hardware_interrupt(co_monitor_t *cmon)
 	 * machine check, NMI and double fault -- none of which should be arriving
 	 * through this path -- but it is a real gap to close alongside the rest.
 	 */
-	func = (void *)(unsigned long)offset;
+	/*
+	 * (size_t), not (unsigned long): Win64 is LLP64 and `unsigned long` is
+	 * four bytes, which would truncate the host ISR's address to its low
+	 * half and jump into nothing. Every Windows handler is above 4 GB.
+	 */
+	func = (void *)(size_t)offset;
 
 	call_intr(func);
 }
